requirements.txt
streamlit
SQLAlchemy
PyMySQL
python-dotenv
passlib[bcrypt]
dateparser
APScheduler
email-validator\
sendgrid
twillio
